# labs/lab03-rest-api/starter/openapi.yaml
openapi: 3.0.3
info:
  title: Lab 3 Item API
  version: 1.0.0
  description: A small REST-style API for managing a collection of items
  description: Each item has an id, a name, and a quantity
  description: Data resets when the server restarts

servers:
  - url: http://localhost:3000

paths:
  /health:
    get:
      summary: Health check
      responses:
        "200":
          description: Server is running
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: ok

  /items:
    get:
      summary: Return all items
      responses:
        "200":
          description: A list of items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Item"
    post:
      summary: Create a new item
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ItemInput"
      responses:
        "201":
          description: The created item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Item"
        "400":
          description: Missing required fields or invalid data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /items/{id}:
    get:
      summary: Return one item by ID
      parameters:
        - $ref: "#/components/parameters/ItemId"
      responses:
        "200":
          description: The matching item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Item"
        "404":
          description: Item not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      summary: Update an existing item
      description: Replaces the name and quantity of the item.
      parameters:
        - $ref: "#/components/parameters/ItemId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ItemInput"
      responses:
        "200":
          description: The updated item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Item"
        "400":
          description: Missing required fields or invalid data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Item not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      summary: Delete an existing item
      parameters:
        - $ref: "#/components/parameters/ItemId"
      responses:
        "204":
          description: Item deleted (no content)
        "404":
          description: Item not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

components:
  parameters:
    ItemId:
      name: id
      in: path
      required: true
      description: The numeric ID of the item.
      schema:
        type: integer
        example: 1

  schemas:
    Item:
      type: object
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: keyboard
        quantity:
          type: integer
          example: 10
      required:
        - id
        - name
        - quantity

    ItemInput:
      type: object
      properties:
        name:
          type: string
          example: monitor
        quantity:
          type: integer
          example: 4
      required:
        - name
        - quantity

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          example: Item not found
      required:
        - error